Move data fetcher form types to CoreBundle. Some fixes after rebase.
